move LoadUser call out of newClient
to start on a fix for user actions when logged in with one user on a
given host with the wf binary, a user should now only be loaded by the
caller of newClient.

newClient no longer takes a bool for authRequired, all calls
updated to new signature and load the user where required.